Solve the cache problem, fixes #6055
I have added a timestamp to the image url, hope this will ensure that no cache image is used by the browser. 
Let us know whether this problem appear again.

However, it is quiet difficult to capture changes on milestones and versions and recreate the image automatically.
Therefore, I added a link at the bottom of the image. So if an image do not change after an update, you can recreate the image manually.